complete invoices routes

+++ app/controllers/invoices_controller.rb
@@ -1,4 +1,5 @@
 class InvoicesController < ApplicationController
+  skip_before_action :verify_authenticity_token
 
     def index
       invoices = []
@@ -23,7 +24,7 @@ class InvoicesController < ApplicationController
 
         invoices.push(construct_invoice)
       end
-      render :json => {
+      render :status => :ok, :json => {
         invoices: invoices
       }
     end
@@ -48,45 +49,42 @@ class InvoicesController < ApplicationController
         construct_invoice['items'] = items
 
       end
-      render :json => {
-        invoices: construct_invoice
+      render :status => :ok, :json => {
+        invoice: construct_invoice
       }
     end
 
     def create
-      invoice = Invoice.create(
-        date: params[:date],
-        supplier_name: params[:supplierName],
-        invoice_number: params[:invoiceNumber],
-        total: params[:total]
+      invoice = Invoice.create!(
+        date: request[:date],
+        supplier_name: request[:supplierName],
+        invoice_number: request[:invoiceNumber],
+        total: request[:total]
       )
-      items = params[:items]
-      p 'items: ', items
-      # items.each do |item|
-      #   Item.create(
-      #     invoice_id: invoice.id,
-      #     code: item.code,
-      #     amount: item.amount
-      #   )
-      # end
-      render :json => {
-        message: 'incomplete create route'
-      }
-    end
-
-    def update
-      render :json => {
-        message: 'incomplete update route'
+      items = request[:items]
+      items.each do |item|
+        Item.create!(
+          code: item['code'],
+          amount: item['amount'],
+          invoice_id: invoice.id
+        )
+      end
+      render :status => :ok, :json => {
+        message: 'Created Invoice and Items.'
       }
     end
 
     def destroy
-      Item.where(invoice_id: params[:id]).each do |item|
-        item.destroy
+      items = Item.where(invoice_id: params[:id])
+      if items
+        items.each do |item|
+          item.destroy
+        end
       end
-      Invoice.find(params[:id]).destroy
-      render :json => {
-        message: 'successful destroy'
+      invoice = Invoice.find(params[:id])
+      invoice.destroy
+      render :status => :ok, :json => {
+        message: 'Destroyed Invoice and Items.'
       }
     end
